Final Version (Draft 4) of Research Topics
This is the final version of my research topic paragraphs. I wasn't entirely sure how long you wanted the length to be. As I didn't want to create too long of a recap of my research, I decided to keep it as brief as I could with all the main points of information I wanted to get across. I am excited to choose a topic and start researching!